split Clesperanto class accross .py file
